ENH: Add two commands get_board_info() and get_board_adc_value_dbg to be available via the A8
get_board_info(): Returns byte1: Board Version, byte2: Board Config
get_board_adc_value_dbg(): If data[0] is 0, returns the board version adc
			   voltage in mV
			 : If data[0] is 1, returns the board config adc
			   voltage in mV

+++ MCU_TASKS/Sources/board_type.c
@@ -35,7 +35,7 @@ uint32_t get_board_adc_value(adc16_chn_t channel){
 *						Each multiple of 0.1 increments the revision 
 *						Returns 0xff on no match
 */
-int get_board_revision(void)
+uint8_t get_board_revision(void)
 {
 	uint32_t board_rev = 0;
 	uint32_t board_adc_val = 0;
